Fix problem when Excel.save with a different filepath doesn't save any data
The original code was doing copy of existing file instead of writing the content - which was causing the changes done in workbook was not saved (in the case in the issue).
So in case of some scenarios it was working, in different was not. Now it works for all scenarios.