Fix reference registered twice: IOStreamnis not closed
The previous implementation of the fix did register the stream before usage because it was not explicitly registered at the processor level. In particular, Argument values were not registered by default when a MetaExpression was assigned to them, while the MetaExpression was registered when in an Assign operator.
Now in CTC-1454, Argument.java:92, agument values (MetaExpressions) are registered by default, so registering again the stream in the IterateConstruct was preventing the closing of the stream when the enclosing MetaExpressionIterator was closed. Removed the now spurious reference registration.