Fix stream closed too early
Registered reference to the stream passed to the IterateConstruct. In the worst case the stream is declared inline, even though it is registered as Argument of the IterateConstruct, it is closed when the provider of the stream is closed, right after it executes, thus preventing iterate from fetching even one element.